requirements.txt
alembic==1.7.4
apparmor==3.0.3
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.4.1
asttokens==2.0.5
autopep8==1.6.0
awscli==1.22.47
awscli-plugin-endpoint==0.4
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
Beaker==1.11.0
black==21.12b0
blinker==1.4
botocore==1.23.47
Brotli==1.0.9
btrfsutil==5.16
CacheControl==0.12.6
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==8.0.3
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
distro==1.6.0
Django==4.0.1
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-dotenv==1.4.2
django-extensions==3.1.5
django-filter==21.1
django-import-export==2.7.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.2.0
docopt==0.6.2
docutils==0.15.2
email-validator==1.1.3
et-xmlfile==1.1.0
executing==0.8.2
extras==1.0.0
filelock==3.6.0
fixtures==3.0.0
Flask==2.0.2
Flask-BabelEx==0.9.4
Flask-Compress==1.8.0
Flask-Gravatar==0.5.0
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-PynamoDB==0.0.2
Flask-Security-Too==4.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gssapi==1.6.14
gufw==21.4.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
ipython==8.0.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
kazam==1.4.5
keyutils==0.6
ldap3==2.9
LibAppArmor==3.0.3
lightdm-gtk-greeter-settings==1.2.2
lit==13.0.0.dev0
Mako==1.1.6
manjaro-sdk==0.1
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-pynamodb==0.8.0
matplotlib-inline==0.1.3
menulibre==2.2.3
meson==0.61.1
more-itertools==8.10.0
msgpack==1.0.3
mugshot==0.4.3
mypy-extensions==0.4.3
npyscreen==4.10.5
odfpy==1.4.1
openpyxl==3.0.9
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.23.2
paramiko==2.9.2
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
pbr==5.8.0
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
pipenv==2022.1.8
platformdirs==2.4.1
ply==3.11
progress==1.6
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyGObject==3.42.0
PyJWT==2.3.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pynamodb==5.2.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PySide6==6.2.3
python-barcode==0.13.1
python-dateutil==2.8.2
python-decouple==3.5
python-distutils-extra==2.39
python-editor==1.0.4
pytz==2021.3
pyxdg==0.27
PyYAML==5.4.1
rados==2.0.0
rbd==2.0.0
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
rsa==4.7.2
s3transfer==0.5.0
shiboken2==5.15.2.1
shiboken6==6.2.3
simplejson==3.17.6
six==1.16.0
speaklater==1.3
SQLAlchemy==1.3.24
sqlparse==0.4.2
sshtunnel==0.4.0
stack-data==0.1.4
tablib==3.1.0
team==1.0
testtools==2.5.0
toml==0.10.2
tomli==2.0.0
traitlets==5.1.1
typing_extensions==4.0.1
udiskie==2.4.1
ufw==0.36.1
urllib3==1.26.8
virtualenv==20.13.3
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==5.3.0
WTForms==2.3.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
